Make expensive operations into cached properties
Accessing the disk or a network resource is very slow and would be
entirely unexpected in a constructor. Cached properties not only avoid
fetching the resources anew each time, but also lets them be loaded
lazily immediately before use. Moving these expensive operations out of
the constructor also adheres to the principle of least surprise.