Store auth information as JSON instead of pickles
Back when I thought that the auth information for Gelbooru may need to
be stored as session cookies, pickles made more sense; however, it has
become clear that only the user ID and password hash need to be stored,
can be serialized in JSON, which is safer.

This change also adds the ability to specify where the authorization
information is saved; by default, it is in the home directory. This has
the side effect of making the code easier to test with test fixtures.