Test and refactor (#3)
* Make expensive operations into cached properties

Accessing the disk or a network resource is very slow and would be
entirely unexpected in a constructor. Cached properties not only avoid
fetching the resources anew each time, but also lets them be loaded
lazily immediately before use. Moving these expensive operations out of
the constructor also adheres to the principle of least surprise.

* Add equality comparison for notes

* Invalidate cached properties after update

* Convert XML to dictionary for easier access

Interfacing with an XML document is cumbersome and annoying. Converting
to a standard Python dictionary not only makes access easier, but also
makes it easier to migrate to JSON-formatted API calls if Gelbooru ever
supports them properly.

* Add first round of tests

* Implement equality comparison for BooruPost

* Add custom exceptions and tests to class discovery

* Make valid site classes into a set

Since the supported sites do not need to be in any particular order,
should be unique, could potentially benefit from a constant-time lookup
and otherwise lack any features unique to lists, it only makes sense to
make it a set.

* Move scale_note() out of the BooruPost class

The function uses fairly little class information and returns a new
Note object, making it ill-suited as an instance method.

* Add CLI tests

* Add coverage.py config

* Rename flake8 config to a standard name

* Add missing dependency

* Update requirements

* Add tox testing

Add tests for all currently-supported versions of Python.

* Change how builtin functions are mocked

This allows the test suite to pass on Python 3.4.

* Update Python version trove classifiers

* Drop Python 2.7 support

Python 2 is an annoying burden and an entirely unnecessary one, given
that this project is designed to be a CLI program, not a library where
broad compatibility is essential. Python 2.7 is still technically
supported until 2020, but its death must come sooner.

* Remove code specific to Python 2

* Store auth information as JSON instead of pickles

Back when I thought that the auth information for Gelbooru may need to
be stored as session cookies, pickles made more sense; however, it has
become clear that only the user ID and password hash need to be stored,
can be serialized in JSON, which is safer.

This change also adds the ability to specify where the authorization
information is saved; by default, it is in the home directory. This has
the side effect of making the code easier to test with test fixtures.